--- lua/base/lsp/formatting.lua
@@ -1,43 +1,0 @@
--- TODO: figure out how to only format on save for certain languages
-return {
-	"stevearc/conform.nvim",
-	event = { "BufReadPre", "BufNewFile" },
-	config = function()
-		local conform = require("conform")
-
-		conform.setup({
-			formatters_by_ft = {
-				go = { "gofumpt", "goimports-reviser" }, -- golines
-				rust = { "rustfmt" },
-				sql = { "sql_formatter" }, -- sqlfluff
-				javascript = { "prettier" },
-				typescript = { "prettier" },
-				javascriptreact = { "prettier" },
-				typescriptreact = { "prettier" },
-				svelte = { "prettier" },
-				css = { "prettier" },
-				html = { "prettier" },
-				json = { "prettier" },
-				yaml = { "prettier" }, -- yamlfix, yamlfmt
-				markdown = { "prettier" },
-				graphql = { "prettier" },
-				lua = { "stylua" },
-				python = { "isort", "black" },
-				proto = { "buf" },
-			},
-			format_on_save = {
-				lsp_fallback = true,
-				async = false,
-				timeout_ms = 500,
-			},
-		})
-
-		vim.keymap.set({ "n", "v" }, "<leader>lf", function()
-			conform.format({
-				lsp_fallback = true,
-				async = false,
-				timeout_ms = 500,
-			})
-		end, { desc = "Format file or range (in visual mode)" })
-	end,
-}
