--- lua/base/lsp/mason.lua
@@ -1,5 +1,0 @@
-return {
-	{
-		"williamboman/mason.nvim",
-	},
-}
