--- lua/base/lsp/treesitter.lua
@@ -1,75 +1,0 @@
-return {
-	{
-		"nvim-treesitter/nvim-treesitter",
-		dependencies = {
-			"nvim-treesitter/nvim-treesitter-textobjects",
-			"JoosepAlviste/nvim-ts-context-commentstring",
-		},
-		build = ":TSUpdate",
-		event = { "BufReadPost", "BufNewFile" },
-		opts = {
-			sync_install = false,
-			ensure_installed = {
-				"bash",
-				"dockerfile",
-				"html",
-				"http",
-				"json",
-				"markdown",
-				"markdown_inline",
-				"org",
-				"query",
-				"regex",
-				"latex",
-				"vim",
-				"vimdoc",
-				"yaml",
-				"sql",
-			},
-			highlight = { enable = true, additional_vim_regex_highlighting = { "org", "markdown" } },
-			indent = { enable = true },
-			ts_context_commentstring = { enable = true, enable_autocmd = false },
-			incremental_selection = {
-				enable = true,
-				keymaps = {
-					init_selection = "gnn",
-					node_incremental = "grn",
-					scope_incremental = "grc",
-					node_decremental = "grm",
-				},
-			},
-			textobjects = {
-				select = {
-					enable = true,
-					lookahead = true, -- Automatically jump forward to textobj, similar to targets.vim
-					keymaps = {
-						-- You can use the capture groups defined in textobjects.scm
-						["aa"] = "@parameter.outer",
-						["ia"] = "@parameter.inner",
-						["af"] = "@function.outer",
-						["if"] = "@function.inner",
-						["ac"] = "@class.outer",
-						["ic"] = "@class.inner",
-					},
-				},
-			},
-			matchup = {
-				enable = true,
-			},
-		},
-		config = function(_, opts)
-			if type(opts.ensure_installed) == "table" then
-				---@type table<string, boolean>
-				local added = {}
-				opts.ensure_installed = vim.tbl_filter(function(lang)
-					if added[lang] then
-						return false
-					end
-					added[lang] = true
-					return true
-				end, opts.ensure_installed)
-			end
-			require("nvim-treesitter.configs").setup(opts)
-		end,
-	},
-}
