+++ lua/config/autocmds.lua
@@ -1,40 +1,3 @@
--- windows to close
-vim.api.nvim_create_autocmd("FileType", {
-	group = augroup("close_with_q"),
-	pattern = {
-		"OverseerForm",
-		"OverseerList",
-		"checkhealth",
-		"floggraph",
-		"fugitive",
-		"git",
-		"help",
-		"lspinfo",
-		"man",
-		"neotest-output",
-		"neotest-summary",
-		"qf",
-		"query",
-		"spectre_panel",
-		"startuptime",
-		"toggleterm",
-		"tsplayground",
-		"vim",
-		"neoai-input",
-		"neotest",
-		"neoai-output",
-		"notify",
-	},
-	callback = function(event)
-		vim.bo[event.buf].buflisted = false
-		vim.keymap.set("n", "q", "<cmd>close<cr>", { buffer = event.buf, silent = true })
-	end,
-})
-
--- TODO: get goto def in new split
--- on_attach = function(client, bufnr)
---   vim.keymap.set('n', 'gs',
---   function()
---     vim.cmd([[ vsplit ]])
---     vim.lsp.buf.definition()
---   end, bufopts)
+-- Autocmds are automatically loaded on the VeryLazy event
+-- Default autocmds that are always set: https://github.com/LazyVim/LazyVim/blob/main/lua/lazyvim/config/autocmds.lua
+-- Add any additional autocmds here
