--- lua/config/globals.lua
@@ -1,7 +1,0 @@
-cmd = vim.cmd -- Command function
-api = vim.api -- Neovim API
-lsp = vim.lsp -- LSP API
-fn = vim.fn -- Vim function
-g = vim.g -- Vim globals
-opt = vim.opt -- Vim optionals
-theme = "catppuccin"
