--- lua/config/lspmaps.lua
@@ -1,11 +1,0 @@
--- this is not used
-local M = {}
-function diagnostic_goto(next, severity)
-	local go = next and vim.diagnostic.goto_next or vim.diagnostic.goto_prev
-	severity = severity and vim.diagnostic.severity[severity] or nil
-	return function()
-		go({ severity = severity })
-	end
-end
-
-return M
