--- lua/plugins/config/signature.lua
@@ -1,5 +1,0 @@
-local sig = require("lsp_signature")
-
-sig.setup({
-	floating_window = true,
-})
