+++ lua/plugins/user.lua
@@ -1,0 +1,200 @@
+local function moons()
+  return [[          ]]
+end
+
+return {
+  {
+    "nvim-telescope/telescope.nvim",
+    keys = {
+      { "<leader>ff", "<cmd>Telescope find_files<cr>", desc = "Find Files" },
+      { "<leader>fr", "<cmd>Telescope oldfiles<cr>", desc = "Recent Files" },
+      { "<leader>fl", "<cmd>Telescope live_grep<cr>", desc = "Grep Search" },
+      { "<leader>fp", "<cmd>Telescope jumplist<cr>", desc = "Find Places" },
+      { "<leader>fg", "<cmd>Telescope git_files<cr>", desc = "Git Files" },
+      { "<leader>fd", "<cmd>Telescope diagnostics<cr>", desc = "Search Diagnostics" },
+      { "<leader>fb", "<cmd>Telescope buffers<cr>", desc = "Buffers" },
+      { "<leader>fh", "<cmd>Telescope help_tags<cr>", desc = "Help" },
+      { "<leader>sa", "<cmd>Telescope autocommands<cr>", desc = "Auto Commands" },
+      { "<leader>sb", "<cmd>Telescope current_buffer_fuzzy_find<cr>", desc = "Buffer" },
+      { "<leader>sc", "<cmd>Telescope command_history<cr>", desc = "Command History" },
+      { "<leader>sC", "<cmd>Telescope commands<cr>", desc = "Commands" },
+      { "<leader>sd", "<cmd>Telescope diagnostics bufnr=0<cr>", desc = "Document diagnostics" },
+      { "<leader>sD", "<cmd>Telescope diagnostics<cr>", desc = "Workspace diagnostics" },
+      { "<leader>sk", "<cmd>Telescope keymaps<cr>", desc = "Key Maps" },
+      { "<leader>sm", "<cmd>Telescope marks<cr>", desc = "Jump to Mark" },
+      { "<leader>so", "<cmd>Telescope vim_options<cr>", desc = "Options" },
+      { "<leader>sR", "<cmd>Telescope resume<cr>", desc = "Resume" },
+      { "<leader>gc", "<cmd>Telescope git_commits<CR>", desc = "commits" },
+      { "<leader>gs", "<cmd>Telescope git_status<CR>", desc = "status" },
+      { "<leader>fP", "<cmd>Telescope projects<CR>", desc = "[F]ind [P]rojects" },
+    },
+    opts = {
+      defaults = {
+        file_ignore_patterns = {
+          "vendor",
+        },
+        mappings = {
+          i = {
+            ["<C-j>"] = function(...)
+              require("telescope.actions").move_selection_next(...)
+            end,
+            ["<C-k>"] = function(...)
+              require("telescope.actions").move_selection_previous(...)
+            end,
+            ["<C-n>"] = function(...)
+              require("telescope.actions").cycle_history_next(...)
+            end,
+            ["<C-p>"] = function(...)
+              require("telescope.actions").cycle_history_prev(...)
+            end,
+          },
+        },
+      },
+    },
+  },
+
+  -- File header
+  {
+    "utilyre/barbecue.nvim",
+    name = "barbecue",
+    version = "*",
+    dependencies = {
+      "SmiteshP/nvim-navic",
+      "nvim-tree/nvim-web-devicons",
+    },
+    opts = {
+      -- configurations go here
+    },
+    -- config = function()
+    -- 	theme = "catppuccin",
+    -- end,
+  },
+  -- Scroll bar
+  {
+    "petertriho/nvim-scrollbar",
+    opts = {},
+    config = {},
+  },
+  -- File Tree
+  {
+    "nvim-neo-tree/neo-tree.nvim",
+    branch = "v2.x",
+    dependencies = {
+      "nvim-lua/plenary.nvim",
+      "MunifTanjim/nui.nvim",
+    },
+    config = function()
+      require("plugins.config.neotree")
+    end,
+  },
+  -- lualine
+  {
+    "nvim-lualine/lualine.nvim",
+    opts = {
+      options = {
+        icons_enabled = true,
+        theme = "catppuccin",
+        section_separators = { left = "", right = "" },
+        component_separators = " ",
+      },
+      sections = {
+        lualine_c = {
+          "%=",
+          { moons },
+        },
+      },
+    },
+  },
+
+  -- Helpers
+  {
+    "rmagatti/goto-preview",
+    config = function()
+      require("plugins.config.goto")
+    end,
+  },
+  {
+    "alexghergh/nvim-tmux-navigation",
+    config = function()
+      require("nvim-tmux-navigation").setup({})
+      vim.keymap.set("n", "<C-h>", "<Cmd>NvimTmuxNavigateLeft<CR>", {})
+      vim.keymap.set("n", "<C-j>", "<Cmd>NvimTmuxNavigateDown<CR>", {})
+      vim.keymap.set("n", "<C-k>", "<Cmd>NvimTmuxNavigateUp<CR>", {})
+      vim.keymap.set("n", "<C-l>", "<Cmd>NvimTmuxNavigateRight<CR>", {})
+    end,
+  },
+  {
+    "chentoast/marks.nvim",
+    config = function()
+      require("marks").setup()
+    end,
+  },
+  {
+    "roobert/surround-ui.nvim",
+    dependencies = {
+      "kylechui/nvim-surround",
+      "folke/which-key.nvim",
+    },
+    config = function()
+      require("surround-ui").setup({
+        root_key = "S",
+      })
+    end,
+  },
+  -- Extras
+  -- found in config/lua/lazy.lua
+
+  -- Theme
+  {
+    "catppuccin/nvim",
+    lazy = true,
+    name = "catppuccin",
+    opts = {
+      integrations = {
+        aerial = true,
+        alpha = true,
+        cmp = true,
+        dashboard = true,
+        flash = true,
+        gitsigns = true,
+        headlines = true,
+        illuminate = true,
+        indent_blankline = { enabled = true },
+        leap = true,
+        lsp_trouble = true,
+        mason = true,
+        markdown = true,
+        mini = true,
+        native_lsp = {
+          enabled = true,
+          underlines = {
+            errors = { "undercurl" },
+            hints = { "undercurl" },
+            warnings = { "undercurl" },
+            information = { "undercurl" },
+          },
+        },
+        navic = { enabled = true, custom_bg = "lualine" },
+        neotest = true,
+        neotree = true,
+        noice = true,
+        notify = true,
+        semantic_tokens = true,
+        telescope = true,
+        treesitter = true,
+        treesitter_context = true,
+        which_key = true,
+      },
+    },
+  },
+  {
+    "LazyVim/LazyVim",
+    opts = {
+      colorscheme = "catppuccin",
+    },
+  },
+
+  -- disable
+  { "akinsho/bufferline.nvim", enabled = false },
+  --{ "nvim-neo-tree/neo-tree.nvim", enabled = false },
+}
